+++ core/src/main/kotlin/org/jetbrains/research/testspark/core/data/LLMWithFeedback.kt
@@ -1,0 +1,13 @@
+package org.jetbrains.research.testspark.core.data
+
+class LLMWithFeedback(val report: Report) {
+
+
+    fun run(){
+        var requestsCount = 0
+        var generatedTestsArePassing = false
+
+
+
+    }
+}
