+++ core/src/main/kotlin/org/jetbrains/research/testspark/core/data/TestCase.kt
@@ -1,0 +1,11 @@
+package org.jetbrains.research.testspark.core.data
+
+/**
+ * Test case storage, implemented based on the org.evosuite.utils.CompactTestCase structure.
+ */
+open class TestCase(
+    val id: Int,
+    var testName: String,
+    var testCode: String,
+    var coveredLines: Set<Int>,
+)
