+++ src/main/kotlin/org/jetbrains/research/testspark/services/ReportLockingService.kt
@@ -4,8 +4,8 @@ import com.intellij.openapi.components.Service
 import com.intellij.openapi.components.service
 import com.intellij.openapi.diagnostic.Logger
 import com.intellij.openapi.project.Project
-import org.jetbrains.research.testspark.data.Report
-import org.jetbrains.research.testspark.data.TestCase
+import org.jetbrains.research.testspark.core.data.Report
+import org.jetbrains.research.testspark.core.data.TestCase
 import org.jetbrains.research.testspark.data.UIContext
 
 @Service(Service.Level.PROJECT)
