+++ src/main/kotlin/org/jetbrains/research/testspark/tools/Manager.kt
@@ -104,68 +104,5 @@ class Manager {
             Llm().generateTestsForLine(project, psiFile, caretOffset, fileUrl, testSamplesCode)
         }
 
-        /**
-         * Displays the given project with the specified number of used tools.
-         * This method is executed asynchronously.
-         *
-         * @param project The project to be displayed.
-         * @param numberOfUsedTool The number of used tools in the project.
-         */
-//        private fun display(project: Project, result: UIContext?, numberOfUsedTool: Int) =
-//            AppExecutorUtil.getAppScheduledExecutorService().execute(Display(project, result, numberOfUsedTool))
     }
 }
-
-/**
- * The Display class represents a display of test generation results in a project.
- * It implements the Runnable interface to run the display in a separate thread.
- *
- * @property project The project in which the display is shown.
- * @property numberOfUsedTool The number of tools used for the test generation.
- */
-//private class Display(private val project: Project, result: UIContext?, private val numberOfUsedTool: Int) : Runnable {
-//    private val log = Logger.getInstance(this::class.java)
-//
-//    override fun run() {
-//        // waiting time after each iteration
-//        val sleepDurationMillis: Long = 1000
-//
-//        // waiting for the generation result
-//        while (true) {
-//            // checks if all generator are finished their work
-//            if (project.service<TestGenerationData>().testGenerationResultList.size != numberOfUsedTool) {
-//                // there is some error during the process running
-//                if (project.service<ErrorService>().isErrorOccurred()) break
-//                log.info("Found ${project.service<TestGenerationData>().testGenerationResultList.size} number of results")
-//                log.info("Waiting for other generation results")
-//                Thread.sleep(sleepDurationMillis)
-//                continue
-//            }
-//
-//            log.info("Found all $numberOfUsedTool generation results")
-//
-//            ApplicationManager.getApplication().invokeLater {
-//                project.service<ReportLockingService>().receiveReport(getMergeResult(numberOfUsedTool))
-//            }
-//
-//            break
-//        }
-//
-//        project.service<RunnerService>().clear()
-//    }
-//
-//    /**
-//     * Retrieves the merged result of a test generation process.
-//     *
-//     * @param numberOfUsedTool The number of tools used for the test generation.
-//     * @return The merged report containing the results of the test generation process.
-//     */
-//    private fun getMergeResult(numberOfUsedTool: Int, generationData: TestGenerationData): Report {
-//        log.info("Merging $numberOfUsedTool generation results")
-//
-//        if (numberOfUsedTool == 1) {
-//            return generationData.testGenerationResultList[0]!!
-//        }
-//        TODO("implement merge")
-//    }
-//}
